cleanup and renames
