+++ test/cbox/makefile
@@ -62,7 +62,7 @@ CFLAGS += -Werror=deprecated-declarations
 CFLAGS += -Wfatal-errors
 
 # Generate dependency files automatically.
-CFLAGS += -g -O0
+CFLAGS += -g -Og
 CFLAGS += -MD -MP -MF $@.d
 CFLAGS += -DDEBUG_BUILD
